fix(runtime): resolve all 55 MSVC compiler warnings
Eliminate every warning emitted by the MSVC build across 19 source files
while preserving full GCC/Clang compatibility for macOS and Linux.

- Wrap #pragma GCC diagnostic blocks with #if !defined(_MSC_VER) guards
- Move block-scope extern declarations to file scope (C4210)
- Add rt_posix_read/rt_posix_write wrappers in rt_file_ext.c for
  size_t-to-unsigned-int narrowing on Windows _read/_write (C4267)
- Add explicit casts for safe narrowing conversions (C4244)
- Replace -(uint64_t)x with 0-(uint64_t)x to avoid C4146
- Use LPCTSTR/LoadCursor instead of LPCWSTR/LoadCursorW (C4133)
- Remove unused variables (C4189), initialize unset pointers (C4701)
- Simplify unreachable loop in rt_regex.c (C4702)
- Use (size_t)1 for Huffman table shift to avoid C4334

+++ src/runtime/graphics/rt_gui_codeeditor.c
@@ -381,7 +381,7 @@ void rt_codeeditor_set_line_number_width(void *editor, int64_t width)
     if (!editor)
         return;
     vg_codeeditor_t *ce = (vg_codeeditor_t *)editor;
-    ce->gutter_width = (int)width * 8; // Approximate char width
+    ce->gutter_width = (float)((int)width * 8); // Approximate char width
 }
 
 void rt_codeeditor_set_gutter_icon(void *editor, int64_t line, void *pixels, int64_t slot)
